// apps/frontend/src/lib/api/generated/schemas/workExperienceResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Resume Genius API
 * OpenAPI spec version: 1.0.0
 */
import type { EmploymentType } from "./employmentType";
import type { WorkExperienceResponseLocation } from "./workExperienceResponseLocation";
import type { WorkExperienceResponseStartDate } from "./workExperienceResponseStartDate";
import type { WorkExperienceResponseEndDate } from "./workExperienceResponseEndDate";
import type { WorkResponsibilityResponse } from "./workResponsibilityResponse";

/**
 * Response model for work experience.
 */
export interface WorkExperienceResponse {
  id: string;
  user_id: string;
  company_name: string;
  position_title: string;
  employment_type: EmploymentType;
  location: WorkExperienceResponseLocation;
  start_date: WorkExperienceResponseStartDate;
  end_date: WorkExperienceResponseEndDate;
  responsibilities: WorkResponsibilityResponse[];
  created_at: string;
  updated_at: string;
}