// apps/frontend/src/lib/api/generated/schemas/workExperienceUpdateRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Resume Genius API
 * OpenAPI spec version: 1.0.0
 */
import type { WorkExperienceUpdateRequestCompanyName } from "./workExperienceUpdateRequestCompanyName";
import type { WorkExperienceUpdateRequestPositionTitle } from "./workExperienceUpdateRequestPositionTitle";
import type { WorkExperienceUpdateRequestEmploymentType } from "./workExperienceUpdateRequestEmploymentType";
import type { WorkExperienceUpdateRequestLocation } from "./workExperienceUpdateRequestLocation";
import type { WorkExperienceUpdateRequestStartDate } from "./workExperienceUpdateRequestStartDate";
import type { WorkExperienceUpdateRequestEndDate } from "./workExperienceUpdateRequestEndDate";

/**
 * Request model for updating work experience.
 */
export interface WorkExperienceUpdateRequest {
  company_name?: WorkExperienceUpdateRequestCompanyName;
  position_title?: WorkExperienceUpdateRequestPositionTitle;
  employment_type?: WorkExperienceUpdateRequestEmploymentType;
  location?: WorkExperienceUpdateRequestLocation;
  start_date?: WorkExperienceUpdateRequestStartDate;
  end_date?: WorkExperienceUpdateRequestEndDate;
}